Create README.md
# penguins_Bot
This bot represents advanced programming in Python, implementing algorithms and data structures to program intelligent agents for a skills programming competition.
There is a game board with a map with a random island order. Each participant (the bot against the enemy) gets a starting island with a number of penguins and must conquer as many islands as possible.
There are different rules and of course we have to evaluate the opponent’s steps to move forward accordingly.
We have developed algorithms that work in response to the agent's actions and future estimation.